Add custom hook  for registerform
